server/requirements.txt
django>=0  # website
djangorestframework>=0  # website
python-dotenv>=0  # .env
psycopg2>=0  # SQL
pymongo>=0  # MongoDB
flake8>=0  # refactor
